+++ R/grey_mode.R
@@ -2,9 +2,7 @@
 #'
 #' @description Grey mode theme with right legend using `greyness` and `linewidthness`.
 #'
-#' @param base_size The base size of the text. Defaults to 11. The title is 110% of this, caption 85% and tag 120%.
-#' @param base_family The family of the text. Defaults to "".
-#' @param ... Provided to support trailing commas only.
+#' @inheritParams light_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -24,15 +22,16 @@
 #'   )
 #'
 grey_mode_r <- function (
-    base_size = 11,
-    base_family = "",
+    text_size = 11,
+    text_family = "",
+    text_colour = greyness[1],
     ...
 ) {
 
   get_mode_r(
-    base_size = base_size,
-    base_family = base_family,
-    text_colour = greyness[1],
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = text_colour,
     axis_line_colour = greyness[2],
     axis_line_linewidth = linewidthness[1],
     panel_line_colour = greyness[3],
@@ -47,7 +46,7 @@ grey_mode_r <- function (
 #'
 #' @description Grey mode theme with top legend using `greyness` and `linewidthness`.
 #'
-#' @inheritParams grey_mode_r
+#' @inheritParams light_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -67,15 +66,16 @@ grey_mode_r <- function (
 #'   )
 #'
 grey_mode_t <- function (
-    base_size = 11,
-    base_family = "",
+    text_size = 11,
+    text_family = "",
+    text_colour = greyness[1],
     ...
 ) {
 
   get_mode_t(
-    base_size = base_size,
-    base_family = base_family,
-    text_colour = greyness[1],
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = text_colour,
     axis_line_colour = greyness[2],
     axis_line_linewidth = linewidthness[1],
     panel_line_colour = greyness[3],
@@ -90,7 +90,7 @@ grey_mode_t <- function (
 #'
 #' @description Grey mode theme with bottom legend using `greyness` and `linewidthness`.
 #'
-#' @inheritParams grey_mode_r
+#' @inheritParams light_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -110,15 +110,16 @@ grey_mode_t <- function (
 #'   )
 #'
 grey_mode_b <- function (
-    base_size = 11,
-    base_family = "",
+    text_size = 11,
+    text_family = "",
+    text_colour = greyness[1],
     ...
 ) {
 
   get_mode_b(
-    base_size = base_size,
-    base_family = base_family,
-    text_colour = greyness[1],
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = text_colour,
     axis_line_colour = greyness[2],
     axis_line_linewidth = linewidthness[1],
     panel_line_colour = greyness[3],
@@ -133,7 +134,7 @@ grey_mode_b <- function (
 #'
 #' @description Grey mode theme with no legend using `greyness` and `linewidthness`.
 #'
-#' @inheritParams grey_mode_r
+#' @inheritParams light_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -153,15 +154,16 @@ grey_mode_b <- function (
 #'   )
 #'
 grey_mode_n <- function (
-    base_size = 11,
-    base_family = "",
+    text_size = 11,
+    text_family = "",
+    text_colour = greyness[1],
     ...
 ) {
 
   get_mode_n(
-    base_size = base_size,
-    base_family = base_family,
-    text_colour = greyness[1],
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = text_colour,
     axis_line_colour = greyness[2],
     axis_line_linewidth = linewidthness[1],
     panel_line_colour = greyness[3],