+++ R/set_blanket.R
@@ -3,7 +3,7 @@
 #' @description Set the default style by setting the default mode and updating a series of geom and annotate defaults.
 #'
 #' @param mode A `*_mode_*` set globally for when mode = NULL. E.g. [light_mode_t()], [grey_mode_r()], or [dark_mode_r()].
-#' @param geom_colour A default geom colour (for geoms other than those used by `annotate_colour`). Fill inherits from this colour. Defaults to `blue`.
+#' @param geom_colour A default geom colour (and fill) for geoms other than those used by `annotate_colour`. Fill inherits from this colour. Defaults to `blue`.
 #' @param annotate_colour A default annotate colour used for `*_vline`, `*_hline`, `*_abline`, `*_curve`, `*_text` and `*_label`. Fill inherits from this colour. Defaults to `lightness[1]`.
 #' @param ... Provided to support trailing commas only.
 #'